requirements.txt
TgCrypto
pyrogram==1.4.16
py-tgcalls
youtube-search-python
python-dotenv
yt-dlp
lyricsgenius
gitpython
requests
aiohttp
aiofiles
pillow
wget
html2text
pyjokes
gTTS
python-arq
gpytranslate
